Light directory refactoring
