+++ UI/portfolio/src/Components/Footer.js
@@ -1,6 +1,6 @@
 import React from "react";
 import styled from "styled-components";
-import NavigationList from "../NavigationList";
+import NavigationList from "./NavigationList";
 
 const FooterSection = styled.div`
   box-sizing: content-box;
