+++ UI/portfolio/src/Components/HeaderComponent/Header.js
@@ -1,5 +1,5 @@
 import React, {useState} from 'react';
-import "./header.css"
+import "./Header.css"
 import ContactStatus from "./contactStatus/ContactButton";
